requirements.txt
ccxt==4.2.5
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
schedule==1.2.0